refactor: typescript GettingStartedTab
