+++ public/locales/en/getStartedTab.yaml
@@ -1,4 +1,5 @@
 TryOut:
+  DocumentationTitle: See full guide
   ButtonText: Try it out!
   title: Confirm
   description: Load a sample save file?
